+++ pages/bills/index.js
@@ -1,0 +1,4 @@
+import Bills from "../../views/bills";
+export default function index() {
+  return <Bills />;
+}
