+++ pages/financialProfile/index.js
@@ -1,4 +1,4 @@
-import FinancialProfile from "../../views/financialProf";
+import FinancialProfile from "../../views/financialProfile";
 export default function index() {
   return <FinancialProfile />;
 }
