--- pages/invoices/index.js
@@ -1,4 +1,0 @@
-import Invoices from "../../views/invoices";
-export default function index() {
-  return <Invoices />;
-}
